Fix modal styles
Previous styles were never checked into the source sass files, so they were unable to be built when we re-built the front page styles.

Instead of adding a selector for each individual modal I changed the selector to the "front-page-popup-container" which all modals except email had. Added front-page-popup-container to the email modal.